pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "simple-sonarqube-api"
version = "0.1.5"
description = "Simple Python client for SonarQube Web API"
readme = "README.md"
requires-python = ">=3.10"
license = "GPL-3.0-only"
license-files = ["LICENSE"]
authors = [{ name = "M.E. (Mikehenrytech)" }]
dependencies = [
  "requests>=2.32"
]

[project.optional-dependencies]
dev = [
  "pytest>=8",
  "responses>=0.25",
  "ruff>=0.6",
  "mypy>=1.11",
  "python-dotenv>=1.2.1",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.bandit]
exclude_dirs = ["tests", ".venv", "build", "dist"]
skips = ["B101"]  # opcional: ignora assert (útil en tests)
confidence = "medium"
severity = "medium"